% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.neEffdecomp.R
\name{ds.neEffdecomp}
\alias{ds.neEffdecomp}
\title{Linear hypotheses for natural effect models}
\usage{
ds.neEffdecomp(model = NULL, datasources = NULL)
}
\arguments{
\item{model}{the name of a fitted natural effect model object. This is the
object saved on the server-side by the \code{ds.neModel} function.}

\item{datasources}{a list of \code{\link{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
a summary table of the object of class c("neLht", "glht") (see glht). The object
additionally inherits from class "neEffdecomp".
}
\description{
This function is similar to R function \code{neEffdecomp} from the 
\code{medflex} package.
}
\details{
The function \code{ds.neEffdecomp} automatically extracts relevant causal
parameter estimates from a natural effect model.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
